requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.0
skia-python~=87.5
